+++ src/runtime_src/xdp/profile/plugin/aie_profile/util/aie_profile_util.h
@@ -55,9 +55,9 @@ namespace xdp::aie::profile {
   };
 
   struct adfAPIResourceInfo {
-    size_t srcPcIdx;
-    size_t destPcIdx;
-    size_t profileResult;
+    uint64_t srcPcIdx;
+    uint64_t destPcIdx;
+    uint64_t profileResult;
     bool isSourceTile = false;
   };
